LayoutTests/imported/w3c/web-platform-tests/selection/fire-selectionchange-event-on-document-if-textcontrol-element-is-in-shadow-tree-expected.txt


FAIL selectionchange event fired on the document in case TextControl element is in Shadow Tree assert_equals: expected 0 but got 1
